Add files via upload
This is an enjoyable Python project idea to emulate the word-guessing game, Hangman. We’ve used a predefined list of words for the guessing aspect, but feel free to improve this by using a third-party dictionary API.

This Python project uses loops, functions, and string formatting to print the hangman’s progress. It also lets you experiment with the standard library's random, time, and os modules.

Specifically, we’ve used the random module to select the word to guess, the os module to clear the screen, and the time module’s .sleep() function to introduce pauses to enhance the game flow.