Add files via upload
Consider a software for maintaining records of the students in a class. Consider the following functions which are required to be performed:

Sorting of names according to First Name of the students.
Finding the Minimum marks among all the marks
Finding contact number of student using his/her First Name.